build: support compilation on IL32P64/LLP64
Windows actually uses IL32P64 instead of the I32LP64 we have
on Linux. This flagged up some conversions in libHX:

	format.c: In function "HXformat2_xvar":
	format.c:476+:40: warning: cast from pointer to integer of
	different size [-Wpointer-to-int-cast]
	tc-map.c: In function "rbt_peel_tree":
	tc-map.c:671:20: warning: cast to pointer from integer of
	different size [-Wint-to-pointer-cast]
	tc-map.c: In function "tmap_rbt_test_7":
	tc-map.c:701:10: warning: cast to pointer from integer of
	different size [-Wint-to-pointer-cast]

+++ src/format.c
@@ -10,6 +10,7 @@
 #include <errno.h>
 #include <stdbool.h>
 #include <stddef.h>
+#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <unistd.h>
@@ -431,7 +432,7 @@ static hxmc_t *HXformat2_xvar(const struct fmt_entry *entry)
 {
 #define IMM(fmt, type) \
 	snprintf(buf, sizeof(buf), (fmt), \
-		static_cast(type, reinterpret_cast(long, entry->ptr))); \
+		static_cast(type, reinterpret_cast(uintptr_t, entry->ptr))); \
 	break;
 #define PTR(fmt, type) \
 	snprintf(buf, sizeof(buf), (fmt), \
